# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Brandyaction OS API
  version: 2.3.0
  description: 지식·업무·목표·콘텐츠·성과를 연결하는 브랜디액션 운영 API
servers:
  - url: https://brandyaction-os.vercel.app
security:
  - bearerAuth: []
paths:
...
  /api/v1/knowledge-documents:
    get:
      summary: 직원 AI가 권한 범위의 문서 원문 조회
      parameters:
        - { in: query, name: organizationId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: documentId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: 문서 원문과 현재 버전 }
        "403": { $ref: "#/components/responses/ApiError" }
    post:
      summary: knowledge.write PAT로 개인 초안 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, title, contentMd]
              properties:
                organizationId: { type: string, format: uuid }
                title: { type: string, minLength: 1, maxLength: 200 }
                contentMd: { type: string, minLength: 1, maxLength: 1500000 }
                folder: { type: string, maxLength: 160, default: AI 저장/검토 대기 }
                tags: { type: array, maxItems: 30, items: { type: string, maxLength: 60 } }
                status: { type: string, enum: [personal_draft, draft], default: personal_draft }
                reason: { type: string, maxLength: 500 }
      responses:
        "201": { description: 생성 문서, documentId와 색인 상태 }
        "429": { $ref: "#/components/responses/ApiError" }
    patch:
      summary: knowledge.write PAT로 버전을 남기고 문서 수정. 새 정본 쓰기 키는 제안을 생성하고 기존 키 동작은 유지
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, documentId]
              properties:
                organizationId: { type: string, format: uuid }
                documentId: { type: string, format: uuid }
                expectedVersion: { type: integer, minimum: 1 }
                title: { type: string, minLength: 1, maxLength: 200 }
                contentMd: { type: string, minLength: 1, maxLength: 1500000 }
                folder: { type: string, maxLength: 160 }
                tags: { type: array, maxItems: 30, items: { type: string, maxLength: 60 } }
                reason: { type: string, maxLength: 500 }
      responses:
        "200": { description: 수정 문서와 새 버전 }
        "202": { description: 정본은 그대로이며 proposalId와 proposal 필드로 변경 제안 반환 }
        "409": { $ref: "#/components/responses/ApiError" }
        "429": { $ref: "#/components/responses/ApiError" }
    delete:
      summary: knowledge.write PAT로 영구 삭제 없이 휴지통 이동
      parameters:
        - { in: query, name: organizationId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: documentId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: reason, schema: { type: string, maxLength: 500 } }
      responses:
        "200": { description: permanent=false인 휴지통 이동 결과 }
        "429": { $ref: "#/components/responses/ApiError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase 사용자 JWT 또는 범위가 지정된 bos_pat 토큰(knowledge.read/knowledge.write/records.read/records.write)
...
  responses:
    ApiError:
      description: 표준 오류 본문
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
                  details: {}